LDlinkR-package {LDlinkR}	R Documentation

LDlinkR: Calculating Linkage Disequilibrium (LD) in Human Population Groups of Interest

Description

Provides access to the 'LDlink' API (https://ldlink.nih.gov/?tab=apiaccess) using the R console. This programmatic access facilitates researchers who are interested in performing batch queries in 1000 Genomes Project (2015) doi: 10.1038/nature15393 data using 'LDlink'. 'LDlink' is an interactive and powerful suite of web-based tools for querying germline variants in human population groups of interest. For more details, please see Machiela et al. (2015) doi: 10.1093/bioinformatics/btv402.

Author(s)

Maintainer: Timothy A. Myers myersta@mail.nih.gov (ORCID)

Authors:

Stephen J. Chanock chanocks@mail.nih.gov

Mitchell J. Machiela mitchell.machiela@nih.gov (ORCID)

See Also

Useful links:

https://ldlink.nih.gov

Report bugs at https://github.com/CBIIT/LDlinkR/issues

[Package LDlinkR version 1.4.0 Index]