LDpair {LDlinkR}	R Documentation

Investigates potentially correlated alleles for a pair of variants.

Description

Investigates potentially correlated alleles for a pair of variants.

Usage

LDpair(
  var1,
  var2,
  pop = "CEU",
  token = NULL,
  output = "table",
  file = FALSE,
  genome_build = "grch37",
  api_root = "https://ldlink.nih.gov/LDlinkRest"
)

Arguments

var1
the first RS number or genomic coordinate (e.g. "chr7:24966446")

var2
the second RS number or genomic coordinate (e.g. "ch7:24966446")

pop
a 1000 Genomes Project population(s), (e.g. YRI or CEU), multiple allowed, default = "CEU"

token
LDlink provided user token, default = NULL, register for token at https://ldlink.nih.gov/?tab=apiaccess

output
two output options available, "text", which displays a two-by-two matrix displaying haplotype counts and allele frequencies along with other statistics, or "table", which displays the same data in rows and columns, default = "table"

file
Optional character string naming a path and file for saving results. If file = FALSE, no file will be generated, default = FALSE.

genome_build
Choose between one of the three options...'grch37' for genome build GRCh37 (hg19), 'grch38' for GRCh38 (hg38), or 'grch38_high_coverage' for GRCh38 High Coverage (hg38) 1000 Genome Project data sets. Default is GRCh37 (hg19).

api_root
Optional alternative root url for API.

Value

text or data frame, depending on the output option

Examples

## Not run: LDpair(var1 = "rs3", var2 = "rs4", pop = "YRI", token = Sys.getenv("LDLINK_TOKEN"))
## Not run: LDpair("rs3", "rs4", "YRI", token = Sys.getenv("LDLINK_TOKEN"), "text")

[Package LDlinkR version 1.4.0 Index]