LDproxy {LDlinkR}	R Documentation

Explore proxy and putative functional variants for a single query variant.

Description

Explore proxy and putative functional variants for a single query variant.

Usage

LDproxy(
  snp,
  pop = "CEU",
  r2d = "r2",
  token = NULL,
  file = FALSE,
  genome_build = "grch37",
  win_size = "500000",
  api_root = "https://ldlink.nih.gov/LDlinkRest"
)

Arguments

snp
an rsID or chromosome coordinate (e.g. "chr7:24966446"), one per query

pop
a 1000 Genomes Project population, (e.g. YRI or CEU), multiple allowed, default = "CEU"

r2d
either "r2" for LD R2 or "d" for LD D', default = "r2"

token
LDlink provided user token, default = NULL, register for token at https://ldlink.nih.gov/?tab=apiaccess

file
Optional character string naming a path and file for saving results. If file = FALSE, no file will be generated, default = FALSE.

genome_build
Choose between one of the three options...'grch37' for genome build GRCh37 (hg19), 'grch38' for GRCh38 (hg38), or 'grch38_high_coverage' for GRCh38 High Coverage (hg38) 1000 Genome Project data sets. Default is GRCh37 (hg19).

win_size
set base pair (bp) window size. Specify a value greater than or equal to zero and less than or equal to 1,000,000bp. Default value is 500,000bp.

api_root
Optional alternative root url for API.

Value

a data frame

Examples

## Not run: LDproxy("rs456", "YRI", "r2", token = Sys.getenv("LDLINK_TOKEN"))

[Package LDlinkR version 1.4.0 Index]