list_chips {LDlinkR}	R Documentation

Provides a data frame listing the names and abbreviation codes for available commercial SNP Chip Arrays from Illumina and Affymetrix.

Description

Provides a data frame listing the names and abbreviation codes for available commercial SNP Chip Arrays from Illumina and Affymetrix.

Usage

list_chips()

Value

a data frame listing the names and abbreviation codes for available SNP Chip Arrays from Illumina and Affymetrix

Examples

list_chips()

[Package LDlinkR version 1.4.0 Index]