list_gtex_tissues {LDlinkR}	R Documentation

Provides a data frame listing the GTEx full names, 'LDexpress' full names (without spaces) and acceptable abbreviation codes of the 54 non-diseased tissue sites collected for the GTEx Portal and used as input for the 'LDexpress' function.

Description

Provides a data frame listing the GTEx full names, 'LDexpress' full names (without spaces) and acceptable abbreviation codes of the 54 non-diseased tissue sites collected for the GTEx Portal and used as input for the 'LDexpress' function.

Usage

list_gtex_tissues()

Value

a data frame listing the GTEx tissues, their names and abbreviation codes used as input for LDexpress.

Examples

list_gtex_tissues()

[Package LDlinkR version 1.4.0 Index]