list_pop {LDlinkR}	R Documentation

Provides a data frame listing the available reference populations from the 1000 Genomes Project.

Description

Provides a data frame listing the available reference populations from the 1000 Genomes Project.

Usage

list_pop()

Value

a data frame listing the available reference populations, continental (ex: European, African, and Admixed American) and sub-populations (ex: Finnish, Gambian, and Peruvian)

Examples

list_pop()

[Package LDlinkR version 1.4.0 Index]